Missed upstream changes

+++ src/device/pci_bridge.c
@@ -584,7 +584,7 @@ pci_bridge_init(const device_t *info)
         interrupt_count = sizeof(interrupts);
         interrupt_mask  = interrupt_count - 1;
         if (dev->slot < 32) {
-            for (int i = 0; i < interrupt_count; i++)
+            for (uint8_t i = 0; i < interrupt_count; i++)
                 interrupts[i] = pci_get_int(dev->slot, PCI_INTA + i);
         }
 
@@ -596,14 +596,14 @@ pci_bridge_init(const device_t *info)
         else
             slot_count = 1; /* AGP bridges always have 1 slot */
 
-    for (int i = 0; i < slot_count; i++) {
+    for (uint8_t i = 0; i < slot_count; i++) {
         /* Interrupts for bridge slots are assigned in round-robin: ABCD, BCDA, CDAB and so on. */
         pci_bridge_log("PCI Bridge %d: downstream slot %02X interrupts %02X %02X %02X %02X\n", dev->bus_index, i, interrupts[i & interrupt_mask], interrupts[(i + 1) & interrupt_mask], interrupts[(i + 2) & interrupt_mask], interrupts[(i + 3) & interrupt_mask]);
         pci_register_bus_slot(dev->bus_index, i, AGP_BRIDGE(dev->local) ? PCI_CARD_AGP : PCI_CARD_NORMAL,
-                                interrupts[i & interrupt_mask],
-                                interrupts[(i + 1) & interrupt_mask],
-                                interrupts[(i + 2) & interrupt_mask],
-                                interrupts[(i + 3) & interrupt_mask]);
+                              interrupts[i & interrupt_mask],
+                              interrupts[(i + 1) & interrupt_mask],
+                              interrupts[(i + 2) & interrupt_mask],
+                              interrupts[(i + 3) & interrupt_mask]);
     }
 
     return dev;
